pyproject.toml
[project]
name = "aqueductus"
version = "1.0.0"
description = ""
authors = [
    {name = "Remo Yukoff",email = "remo.yukoff@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click (>=8.1.8,<9.0.0)",
    "boto3 (>=1.36.13,<2.0.0)",
    "pyathena (>=3.12.2,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.0"
flake8 = "^7.1.1"
flake8-bugbear = "^24.12.12"
pytest = "^8.3.4"

[tool.isort]
profile = "black"

[tool.poetry.scripts]
aqueductus = "aqueductus.main:main"